todo/requirements.txt
aniso8601==7.0.0
asgiref==3.5.2
automium==0.2.6
automium-web==0.1.1
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.15
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
Faker==12.0.1
Flask==2.2.1
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.3
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
legacy==0.1.6
Markdown==3.4.1
MarkupSafe==2.1.1
mixer==7.2.2
packaging==21.3
promise==2.3
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
uritemplate==4.1.1
urllib3==1.26.12
users==1.0.dev0
views==0.3
Werkzeug==2.2.1
